fixes the display issue for the trend charts
This does need a new chart system in the future.
