+++ app/Http/Livewire/Trends/DailyAverage.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Livewire\Trends;
+
+use Livewire\Component;
+
+class DailyAverage extends Component
+{
+    public $trendDailyAvg;
+    public $trendDailyAvgLabels;
+
+    public function render()
+    {
+        return view('livewire.trends.daily-average');
+    }
+}
